Moved notification trigger into a separate class
