+++ apps/onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/mapper/NotificationMapper.java
@@ -17,4 +17,5 @@ public interface NotificationMapper {
     @Mapping(source = "productId", target = "product")
     @Mapping(source = "institution.id", target = "internalIstitutionID")
     NotificationToSend toSCContracts(Onboarding onboarding);
+
 }
