+++ apps/onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/service/NotificationEventService.java
@@ -2,7 +2,7 @@ package it.pagopa.selfcare.onboarding.service;
 
 import it.pagopa.selfcare.onboarding.entity.Onboarding;
 
-public interface MessageService {
+public interface NotificationEventService {
 
     void send(Onboarding onboarding);
 }
